addStudent의 GPA를 float형에서 double형으로 바꿈

+++ src/main/java/com/ajou/TwotoSix/unitTest2/service/MockService.java
@@ -25,7 +25,7 @@ public class MockService {
         return student.getGPA();
     }
   
-    public Student addStudent(String name, String studentId, int currentSemester, String major, float GPA){
+    public Student addStudent(String name, String studentId, int currentSemester, String major, double GPA){
         Student addStudent = new Student(name, studentId, currentSemester, major, GPA);
         return addStudent;
     }
